Refactor OpenStack actions to match AWS.
I went through all of the actions and compared them against the AWS
equivalients. There are a few subtle differences in the instance
actions, but for the most part they should match up. For the OpenStack
API the concept of 'pausing' and 'unpausing' stores the machine state in
memory instead of on disk.

The functionality in this commit should encompass the following
Vagrant commands against the OpenStack API:
        - vagrant destroy <vm>
        - vagrant up <vm>
        - vagrant reload <vm>
        - vagrant provision <vm>
        - vagrant halt <vm>
        - vagrant suspend <vm>